shuffle code around to remove import loop
