+++ cmd/kubermatic-api/main.go
@@ -220,9 +220,9 @@ func createInitProviders(ctx context.Context, options serverRunOptions, masterCf
 
 	var configGetter provider.KubermaticConfigurationGetter
 	if options.kubermaticConfiguration != nil {
-		configGetter, err = provider.StaticKubermaticConfigurationGetterFactory(options.kubermaticConfiguration)
+		configGetter, err = kubernetesprovider.StaticKubermaticConfigurationGetterFactory(options.kubermaticConfiguration)
 	} else {
-		configGetter, err = provider.DynamicKubermaticConfigurationGetterFactory(client, options.namespace)
+		configGetter, err = kubernetesprovider.DynamicKubermaticConfigurationGetterFactory(client, options.namespace)
 	}
 	if err != nil {
 		return providers{}, err
@@ -244,7 +244,7 @@ func createInitProviders(ctx context.Context, options serverRunOptions, masterCf
 		return providers{}, errors.New("failed to sync mgr cache")
 	}
 
-	seedClientGetter := provider.SeedClientGetterFactory(seedKubeconfigGetter)
+	seedClientGetter := kubernetesprovider.SeedClientGetterFactory(seedKubeconfigGetter)
 	clusterProviderGetter := clusterProviderFactory(mgr.GetRESTMapper(), seedKubeconfigGetter, seedClientGetter, options)
 
 	presetProvider, err := kubernetesprovider.NewPresetProvider(client)
