+++ pkg/defaulting/cluster.go
@@ -24,6 +24,7 @@ import (
 	"go.uber.org/zap"
 
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
+	kubermaticv1helper "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1/helper"
 	"k8c.io/kubermatic/v2/pkg/provider"
 	"k8c.io/kubermatic/v2/pkg/resources"
 	"k8c.io/kubermatic/v2/pkg/version/cni"
@@ -105,7 +106,7 @@ func DefaultClusterSpec(ctx context.Context, spec *kubermaticv1.ClusterSpec, tem
 	}
 
 	// Ensure provider name matches the given spec
-	providerName, err := provider.ClusterCloudProviderName(spec.Cloud)
+	providerName, err := kubermaticv1helper.ClusterCloudProviderName(spec.Cloud)
 	if err != nil {
 		return fmt.Errorf("failed to determine cloud provider: %w", err)
 	}
