+++ pkg/ee/constraint-template-controller/util.go
@@ -29,7 +29,7 @@ import (
 	"fmt"
 
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
-	"k8c.io/kubermatic/v2/pkg/provider"
+	kubermaticv1helper "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1/helper"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -68,7 +68,7 @@ func filterProvider(ct *kubermaticv1.ConstraintTemplate, ctList []kubermaticv1.C
 	providersSet := sets.NewString(ct.Spec.Selector.Providers...)
 
 	for _, cluster := range ctList {
-		name, err := provider.ClusterCloudProviderName(cluster.Spec.Cloud)
+		name, err := kubermaticv1helper.ClusterCloudProviderName(cluster.Spec.Cloud)
 		if err == nil && providersSet.Has(name) {
 			filteredList = append(filteredList, cluster)
 		}
