+++ pkg/handler/common/cluster.go
@@ -1139,7 +1139,7 @@ func ValidateClusterSpec(updateManager common.UpdateManager, body apiv1.CreateCl
 		return errors.New("invalid cluster name: too long (greater than 100 characters)")
 	}
 
-	providerName, err := provider.ClusterCloudProviderName(body.Cluster.Spec.Cloud)
+	providerName, err := kubermaticv1helper.ClusterCloudProviderName(body.Cluster.Spec.Cloud)
 	if err != nil {
 		return fmt.Errorf("failed to get the cloud provider name: %w", err)
 	}
