+++ pkg/resources/resources.go
@@ -480,6 +480,9 @@ const (
 	KubeVirtInfraSecretName = "cloud-controller-manager-infra-kubeconfig"
 	// KubeVirtInfraSecretKey infra kubeconfig.
 	KubeVirtInfraSecretKey = "infra-kubeconfig"
+
+	// DefaultNodePortRange is a Kubernetes cluster's default nodeport range.
+	DefaultNodePortRange = "30000-32767"
 )
 
 const (
