+++ pkg/validation/cluster.go
@@ -27,6 +27,7 @@ import (
 	"github.com/coreos/locksmith/pkg/timeutil"
 
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
+	kubermaticv1helper "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1/helper"
 	"k8c.io/kubermatic/v2/pkg/features"
 	kuberneteshelper "k8c.io/kubermatic/v2/pkg/kubernetes"
 	"k8c.io/kubermatic/v2/pkg/provider"
@@ -350,7 +351,7 @@ func ValidateClusterNetworkConfig(n *kubermaticv1.ClusterNetworkingConfig, dc *k
 	}
 
 	if n.IPFamily == kubermaticv1.IPFamilyDualStack && dc != nil {
-		cloudProvider, err := provider.DatacenterCloudProviderName(&dc.Spec)
+		cloudProvider, err := kubermaticv1helper.DatacenterCloudProviderName(&dc.Spec)
 		if err != nil {
 			allErrs = append(allErrs, field.Invalid(fldPath, nil,
 				fmt.Sprintf("could not determine cloud provider: %v", err)))
@@ -532,12 +533,12 @@ func ValidateCloudChange(newSpec, oldSpec kubermaticv1.CloudSpec) error {
 		return errors.New("changing the datacenter is not allowed")
 	}
 
-	oldCloudProvider, err := provider.ClusterCloudProviderName(oldSpec)
+	oldCloudProvider, err := kubermaticv1helper.ClusterCloudProviderName(oldSpec)
 	if err != nil {
 		return fmt.Errorf("could not determine old cloud provider: %w", err)
 	}
 
-	newCloudProvider, err := provider.ClusterCloudProviderName(newSpec)
+	newCloudProvider, err := kubermaticv1helper.ClusterCloudProviderName(newSpec)
 	if err != nil {
 		return fmt.Errorf("could not determine new cloud provider: %w", err)
 	}
@@ -559,7 +560,7 @@ func ValidateCloudSpec(spec kubermaticv1.CloudSpec, dc *kubermaticv1.Datacenter,
 		allErrs = append(allErrs, field.Required(parentFieldPath.Child("dc"), "no node datacenter specified"))
 	}
 
-	providerName, err := provider.ClusterCloudProviderName(spec)
+	providerName, err := kubermaticv1helper.ClusterCloudProviderName(spec)
 	if err != nil {
 		allErrs = append(allErrs, field.Invalid(parentFieldPath, "<redacted>", err.Error()))
 	}
@@ -574,12 +575,12 @@ func ValidateCloudSpec(spec kubermaticv1.CloudSpec, dc *kubermaticv1.Datacenter,
 	}
 
 	if dc != nil {
-		clusterCloudProvider, err := provider.ClusterCloudProviderName(spec)
+		clusterCloudProvider, err := kubermaticv1helper.ClusterCloudProviderName(spec)
 		if err != nil {
 			allErrs = append(allErrs, field.Invalid(parentFieldPath, nil, fmt.Sprintf("could not determine cluster cloud provider: %v", err)))
 		}
 
-		dcCloudProvider, err := provider.DatacenterCloudProviderName(&dc.Spec)
+		dcCloudProvider, err := kubermaticv1helper.DatacenterCloudProviderName(&dc.Spec)
 		if err != nil {
 			allErrs = append(allErrs, field.Invalid(parentFieldPath, nil, fmt.Sprintf("could not determine datacenter cloud provider: %v", err)))
 		}
